Adding support to REST API
Adding the support for REST API, in the new request the user can
access the data and send requests by html protocol.

To send request for relay a POST request must be done in recebe.php file.

To get all values of temperature, a POST request must be done in
index.php file. The same for humidity in umidade.php file.

To get the current temperature and humidity, a request must be
done in temp.php file.

Signed-off-by: Robert Cabral <robertcabral.qx@gmail.com>